fix infinite wait for config update when call to App Store fails
Play Store asks the GmsCompat app for updates of GmsCompatConfig before updating Play services or
Play Store to avoid the update failing due to outdated max support version constraint (which is
specified in GmsCompatConfig). In response, GmsCompat app asks the App Store to update
GmsCompatConfig and waits for the response from App Store.

When the App Store is not available (e.g. when it's disabled), call to the App Store fails and both
the calling thread in GmsCompat app and the calling thread inside Play Store remained stalled due to
the missing callback.